# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CircumventionRequest": {
        "properties": {
          "country": {
            "default": "",
            "maxLength": 2,
            "title": "Country",
            "type": "string"
          },
          "custom_bridges": {
            "items": {
              "type": "string"
            },
            "maxItems": 40,
            "title": "Custom Bridges",
            "type": "array"
          },
          "mode": {
            "pattern": "^(direct|auto|manual)$",
            "title": "Mode",
            "type": "string"
          },
          "transport": {
            "default": "snowflake",
            "maxLength": 32,
            "title": "Transport",
            "type": "string"
          }
        },
        "required": [
          "mode"
        ],
        "title": "CircumventionRequest",
        "type": "object"
      },
...
      "DeviceAccessRequest": {
        "properties": {
          "alias": {
            "default": "",
            "maxLength": 64,
            "title": "Alias",
            "type": "string"
          },
          "mac": {
            "maxLength": 17,
            "minLength": 11,
            "title": "Mac",
            "type": "string"
          },
          "schedule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceScheduleRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "mac"
        ],
        "title": "DeviceAccessRequest",
        "type": "object"
      },
...
      "DevicePauseRequest": {
        "properties": {
          "mac": {
            "maxLength": 17,
            "minLength": 11,
            "title": "Mac",
            "type": "string"
          },
          "minutes": {
            "maximum": 1440.0,
            "minimum": 0.0,
            "title": "Minutes",
            "type": "integer"
          }
        },
        "required": [
          "mac",
          "minutes"
        ],
        "title": "DevicePauseRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OnionPi API",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/circumvention": {
      "get": {
        "operationId": "circumvention_state_api_v1_circumvention_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Circumvention State Api V1 Circumvention Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Circumvention State"
      },
      "post": {
        "operationId": "update_circumvention_api_v1_circumvention_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CircumventionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Circumvention Api V1 Circumvention Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Circumvention"
      }
    },
    "/api/v1/circumvention/refresh": {
      "post": {
        "description": "Updates the built-in bridge list straight from the Tor Project.\n\nRequires a working direct connection to bridges.torproject.org, which is\nexactly what censorship removes: failure here is expected, not fatal.",
        "operationId": "refresh_circumvention_api_v1_circumvention_refresh_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Refresh Circumvention Api V1 Circumvention Refresh Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Circumvention"
      }
    },
...
    "/api/v1/devices/access": {
      "get": {
        "operationId": "device_access_api_v1_devices_access_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Device Access Api V1 Devices Access Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Device Access"
      },
      "post": {
        "operationId": "update_device_access_api_v1_devices_access_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceAccessRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Device Access Api V1 Devices Access Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Device Access"
      }
    },
    "/api/v1/devices/access/pause": {
      "post": {
        "operationId": "pause_device_api_v1_devices_access_pause_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevicePauseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Pause Device Api V1 Devices Access Pause Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Device"
      }
    },
    "/api/v1/devices/access/remove": {
      "post": {
        "operationId": "remove_device_access_api_v1_devices_access_remove_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceMacRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Remove Device Access Api V1 Devices Access Remove Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Device Access"
      }
    },
    "/api/v1/devices/block": {
      "post": {
        "operationId": "block_device_api_v1_devices_block_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceBlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Block Device Api V1 Devices Block Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Block Device"
      }
    },
    "/api/v1/dns-filter": {
      "get": {
        "operationId": "dns_filter_state_api_v1_dns_filter_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dns Filter State Api V1 Dns Filter Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Dns Filter State"
      },
      "post": {
        "operationId": "update_dns_filter_api_v1_dns_filter_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DnsFilterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Update Dns Filter Api V1 Dns Filter Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Dns Filter"
      }
    },
    "/api/v1/dns-filter/refresh": {
      "post": {
        "operationId": "refresh_dns_filter_api_v1_dns_filter_refresh_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Refresh Dns Filter Api V1 Dns Filter Refresh Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Dns Filter"
      }
    },
...
    "/api/v1/onion": {
      "post": {
        "operationId": "set_onion_api_v1_onion_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Set Onion Api V1 Onion Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Onion"
      }
    },
    "/api/v1/onion/clients": {
      "post": {
        "description": "Authorises one device. The returned key is never stored in clear.",
        "operationId": "add_onion_client_api_v1_onion_clients_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnionClientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Onion Client Api V1 Onion Clients Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Onion Client"
      }
    },
    "/api/v1/onion/clients/remove": {
      "post": {
        "operationId": "remove_onion_client_api_v1_onion_clients_remove_post",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnionClientRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Remove Onion Client Api V1 Onion Clients Remove Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Onion Client"
      }
    },
...
  }
}